refactor(eth): move provider into sdk/eth and drop dead multicall from runtime
The eth `getProvider`/`DummyProvider`/`QueuedStaticJsonRpcProvider` lived in
`packages/runtime` even though the runtime never used them — every consumer is
in `packages/sdk/src/eth`. Every other chain builds its client in
`packages/sdk/src/<chain>/network.ts`; eth was the lone exception.

- Move `provider.ts` (+ its test) to `packages/sdk/src/eth/`, importing
  `Endpoints`/metrics/`GLOBAL_CONFIG` from `@sentio/runtime`.
- Point `eth/index.ts`, `generic-processor.ts`, `eth-plugin.ts`, and
  `utils/block.test.ts` at the new location. Generated eth processors are
  unaffected (they import via the `@sentio/sdk` barrel).
- Delete `runtime/src/multicall.ts`: it was never exported from the runtime
  index nor imported anywhere (dead code).
- Drop `ethers` and `lru-cache` from `packages/runtime` deps — with provider
  and multicall gone, runtime has no remaining `ethers` references, decoupling
  the engine from a chain-specific (eth) dependency.
