+++ packages/runtime/package.json
@@ -41,10 +41,8 @@
     "@opentelemetry/sdk-node": "^0.218.0",
     "@sentio/protos": "workspace:*",
     "commander": "^14.0.0",
-    "ethers": "npm:@sentio/ethers@6.13.1-patch.6",
     "fs-extra": "^11.2.0",
     "ix": "^7.0.0",
-    "lru-cache": "^11.1.1",
     "piscina": "5.1.3",
     "rxjs": "^7.8.1",
     "utility-types": "^3.11.0",
