+++ packages/runtime/src/index.ts
@@ -6,5 +6,4 @@ export * from './chain-config.js'
 export * from './service-v3.js'
 export { GLOBAL_CONFIG, type GlobalConfig } from './global-config.js'
 export * from './db-context.js'
-export * from './provider.js'
 export * from './metrics.js'
