+++ packages/sdk/src/eth/tests/provider.test.ts
@@ -1,10 +1,9 @@
 import assert from 'assert'
 import { describe, test, mock } from 'node:test'
 import { JsonRpcProvider } from 'ethers'
-import { getProvider } from './provider.js'
-import { Endpoints } from './endpoints.js'
+import { getProvider } from '../provider.js'
+import { Endpoints, GLOBAL_CONFIG } from '@sentio/runtime'
 import { EthChainId } from '@sentio/chain'
-import { GLOBAL_CONFIG } from './global-config.js'
 
 Endpoints.INSTANCE.concurrency = 1
 Endpoints.INSTANCE.batchCount = 1
